runtime: prototype CL showing possible implementation of arenas
This is a sample implementation of arenas in Go for amd64. Since this is
a prototype, it may have various issues and limitations, and not quite
follow the proposal.

Also, the setup of the arena package, and its interaction with the
reflect and runtime packages is just one possibility and seems in need
of cleanup. It is not yet clear yet what the best arrangement is, given
that the arena implementation needs to use some private reflect data
structures.

DO NOT SUBMIT
DO NOT REVIEW

Change-Id: I27f78027744e99cf385218375036db326cda70bd